revert to SpringBoot 3.1.X

+++ src/main/java/it/gov/pagopa/initiative/connector/error/decoder/Custom5xxErrorDecoder.java
@@ -24,7 +24,7 @@ public class Custom5xxErrorDecoder implements ErrorDecoder {
                     exception.getMessage(),
                     response.request().httpMethod(),
                     exception,
-                    (Long)null,
+                    null,
                     response.request());
         }
         return exception;
